+++ model_navigator/framework_api/commands/config_cli.py
@@ -11,13 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 from pathlib import Path
-from typing import Dict, Optional, Tuple
+from typing import Dict, List, Optional, Tuple
 
-from model_navigator.cli.convert_model import ConversionSetConfig
-from model_navigator.converter.config import TensorRTPrecision
+from model_navigator.converter.config import ComparatorConfig, TensorRTPrecision
 from model_navigator.framework_api.commands.core import Command, CommandType
+from model_navigator.framework_api.status import RuntimeResults
 from model_navigator.framework_api.utils import (
     Extension,
     Framework,
@@ -43,21 +42,6 @@ def extension2format(extension: str):
         return None
 
 
-TRITON_SUPPORTED_FORMATS_PYT = [
-    Format.ONNX,
-    Format.TENSORRT,
-    Format.TORCHSCRIPT,
-    Format.TORCH_TRT,
-]
-
-TRITON_SUPPORTED_FORMATS_TF = [
-    Format.TF_TRT,
-    Format.TF_SAVEDMODEL,
-    Format.ONNX,
-    Format.TENSORRT,
-]
-
-
 class ConfigCli(Command):
     supported_extensions = [Extension.ONNX.value, Extension.PT.value, Extension.SAVEDMODEL.value]
 
@@ -65,6 +49,9 @@ class ConfigCli(Command):
         self,
         name: str,
         target_format: Format,
+        runtime_results: List[RuntimeResults],
+        atol: Dict[str, float],
+        rtol: Dict[str, float],
         requires: Tuple[Command, ...] = (),
         target_jit_type: Optional[JitType] = None,
         target_precision: Optional[TensorRTPrecision] = None,
@@ -72,6 +59,9 @@ class ConfigCli(Command):
         super().__init__(name=name, command_type=CommandType.GEN_CONFIG, requires=requires, target_format=target_format)
         self.target_jit_type = target_jit_type
         self.target_precision = target_precision
+        self.runtime_results = runtime_results
+        self.atol = atol
+        self.rtol = rtol
 
     def __call__(
         self,
@@ -100,22 +90,11 @@ class ConfigCli(Command):
                 inputs=input_metadata,
                 outputs=output_metadata,
             )
-            conversion_set_config = ConversionSetConfig(
-                onnx_opsets=[opset],
-                target_formats=TRITON_SUPPORTED_FORMATS_PYT
-                if framework == Framework.PYT
-                else TRITON_SUPPORTED_FORMATS_TF,
-                tensorrt_precisions=[self.target_precision]
-                if self.target_precision
-                else [TensorRTPrecision.FP16, TensorRTPrecision.TF32],
-            )
-            # comparator_config = ComparatorConfig()
-            # dataset_profile_config = DatasetProfileConfig()
+
+            comparator_config = ComparatorConfig(atol=self.atol, rtol=self.rtol)
 
             config_file.save_config(src_model_config)
             config_file.save_config(model_signature_config)
-            config_file.save_config(conversion_set_config)
-            # config_file.save_config(comparator_config)
-            # config_file.save_config(dataset_profile_config)
+            config_file.save_config(comparator_config)
 
         return config_relative_path